% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibration}
\alias{calibration}
\title{Calibration of an image based on the scale bar}
\usage{
calibration(x, cal.lenght)
}
\arguments{
\item{x}{image of class \code{magick-image} or a local path to image or URL}

\item{cal.length}{length or scale bar}
}
\value{
numeric of pixel values corresponding to the cal.length
}
\description{
Calibration of an image based on the scale bar
}
