% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_from_image.R
\name{coords_from_image}
\alias{coords_from_image}
\title{Sample coordinates from image}
\usage{
coords_from_image(x, n)
}
\arguments{
\item{x}{is of class cimg}

\item{n}{is the number of coordinates (positive integer)}
}
\value{
coordinates x and y
}
\description{
Sample coordinates from image
}
\examples{
a.m <- load.image("https://upload.wikimedia.org/wikipedia/commons/0/02/2006-10-25_Amanita_muscaria_crop.jpg")
xy <- coords_from_image(a.m, n = 2)

}
