% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\title{Measure objects in images}
\usage{
measure(x, cal)
}
\arguments{
\item{x}{image of class \code{magick-image} or a local path to image or URL}

\item{cal}{is a calibration numeric of pixel corresponding to a scale bar in the image. Usually output from \link{calibration}}
}
\description{
Measure objects in images
}
